Various Quality of Life Changes
- Updated the profile page, adding a save bio button and changing
behavior of updating a profile. Updated the post and get profile
routes to allow sending of bio and image information. Added
configuration to the page to dynamically style and display elements
based on if the user if viewing their private profile or a public
profile.

- Updated the CSS throughout website and basic HTML structure to
be more standardized. Moved market css to its own file.