Added Market-Listing Insert and Query functions to MongoDB and improves state in market.js
Two POST routes "/get-listings" and "/post-listing" have been added to route-handler.js, account-routes.js, and accounts.js
"/get-listings" takes a body that represents a query that the database returns through a response
"/post-listing" takes a body that represents a listing that the query inserts into the correct collection

mongodb-library.js was modified so that it's get_data function can take optional, defaulted arguments to allows sorting, whie functioning the same way as before when used in the old way without sort (done this way for backwards compatibility)

market.js was reworked to fix a few bugs regarding state and filtering.
Now, all states used for filtering are maintained in the "main" component.
Also, a few bugs were fixed regarding the persistence of filter options after the filter window is closed
Finally, data validation was improved and pulled up to parent components in some components